Extracting view functionality to separate files
